refactor(windows): extract 'network_changes' module
